app/requirement.txt
# Core Libraries
streamlit==1.35.0
python-dotenv==1.0.1
PyPDF2==3.0.1

# LangChain & Vectorstore
langchain==0.3.27
langchain-community==0.0.36
faiss-cpu==1.7.4

# Embeddings
sentence-transformers==2.6.1
transformers==4.41.1
torch==2.3.0

# OpenAI (native)
openai==1.33.0

# HuggingFace Instructor Embeddings
InstructorEmbedding==1.0.1

